Improve tests and logging.
Log the actual query made to postgres (sqlalchemy was logging a template of the query before).

Don't ask for a needless test file.

Change the slow query test (used to see if a query can be canceled) to use a query which should be slower.

Fix a few style issues.