Database assetstore.
The database assetstore supports Mongo and sqlalchemy SQL types (provided dependencies are installed).  For postgres, all non-volatile functions are allowed (as for db items).  Each file in the assetstore references a specific table or collection.  The download endpoint uses extraParameters to perform full selects on the database (with the same basic parameters as used by the item select endpoint).  Default values for sort, fields, filters, limit, and format can be specified, and these can be overridden in the query.

To support range-requests, we preload the response to calculate its length.  This is inefficient and a waste of memory.

The mongo database connector now caches field information.  Since it inspects the entire collection to determine available fields, this is an expensive operation.  This could still use improvement.

If _id is not explicitly included in a field list for Mongo, it is excluded, so as to only return the fields that were actually specified.

Tests are needed.