Updated import data to work with schemata and multiple databases.
Some database connectors (mongo) support a generic connection that can handle more than one database.  For these, the connection string can leave out the database, and then the database must be specified as part of the file database information.  sqlalchemy doesn't support enumerating databases, so, for now, this isn's supported for sql-based connections (failure to include the database throws an error).

Some databases have compound table names (schema plus table name).  This is now supported.

Enumerating tables will also enumerate views.  Tables and views from all reachable schemata are included.  Some schemata probably shouldn't be listed as first-class items, since postgres exposes them differently, so for now this is more inclusive rather than strictly necessary.

As a consequence, the database_assetstore endpoints have changed in format to include additional information, and the web client has been changed to match.

You can now add tables to collections and users.  If so done, a folder is created with the database name.  Such folders are also created for multiple-database connections.

When dumping data to JSON, mongo connections now use the bson module, as data from mongo may not be convertible to strings using the standard json module.  This means that timestamps are less readable, but preserve data type better.

When dumping data to CSV, if the data type is not simple, or is a bson.binary.Binary datatype, it is stringified with JSON prior to converting to CSV.  Perhaps the binary data shouldn't be so cast (as it can be placed directly into the csv), but this could be changed if desired.  Perhaps csv needs a dialect option.