Add more functions for pythonic use of database assetstores.
Expose getTableList as a method of an assetstore adapter and as a globally available function.  Make it easy to exclude internal tables.

Add the ability to get a connector to a table without having a file.  Add methods to make it easier to query such a table.

Add raw output formats so that the results can be consumed by a python client.

Return items that were imported from the import function.