Add first-order MySQL support.
This adds support for regex searches and MySQL functions.

Tests are performed against a public database.  The table used has been placed on data.kitware.com, so if the public MySQL server goes down, we can switch to using a local server.

mysqlclient has been added to the requirements.txt.  It is not strictly necessary in MySQL support is not required, and alternate mysql python modules could be used instead.