Add explicit support for SQLite.
This allows sqlite to reference a file on the local file system or any file in a Girder filesystem assetstore.  For files in a Girder assetstore, the connection uri is of the form: `sqlite:///<resource path of Girder file>`, for instance `sqlite:////user/manthey/Public/testdb/testdb`.  SQLite supports functions, regex and match (search).  It has complete tests.