[WIP] Allow users to create database files.
Before, to create a file you first had to create an assetstore, then import a file from it.  This provides settings where, if enabled, users may create files directly as an alternative to uploading a file in the traditional manner.

This adds an assetstore that is used for all user-created database items.  This assetstore always has the same ID, and is hidden on the assetstore web page when the plugin is enabled.  If the plugin is disabled, this assetstore is removed.  It will appear in the assetstore list if the plugin is manually removed or disabled, at which point it can be deleted manually.

Added a settings page:
  - enable user file creation
  - allow specific groups to be restricted to specific database URIs.  This can be used to limit the hosts, database types, etc.

TODO:
- Add button on upload dialog to switch to the create file dialog
- Add create file dialog.  This is a super-set of the import functionality which also lists a database uri, but does not need to have the destination fields
- Add a button to check access to the database uri (both in the file dialog and the create assetstore panel)
- Add a query builder to help construct fields, sort, filters.  This would not support general functions at this time.
- Show the current default limit (try to report what it might be based on the current data with the current filters)