Remove the requirement of dbtype.
Before, we asked the user for a dbtype, would store it, and make a
canonical dialect value as well.  This is a needless burden to the user,
since the connection string (URI) is sufficient, and if we add
additional database types, we would not automatically switch to them.
Instead, we save an initial database type, add a special code path for
accessing the generic sqlalchemy type, and otherwise recheck the
database type when we need it.